+++ src/app/components/supervisorToolbar/supervisorToolbarController.js
@@ -4,7 +4,6 @@ angular.module('liveopsConfigPanel')
   .controller('supervisorToolbarController', ['$scope', '$sce', '$location', 'config2Url',
     function ($scope, $sce, $location, config2Url) {
       $scope.supervisorToolbarHostname = $sce.trustAsResourceUrl(config2Url + '/#/supervisorToolbar');
-      $scope.supervisorToolbarIsVisible = $location.search()['alpha'];
       $scope.$watch(function(){
         return $location.path();
       }, function(value){